rpcss: Constitution of Core Collections by Principal Component Scoring Strategy

Generate a Core Collection with Principal Component Scoring Strategy (PCSS) using qualitative and/or quantitative trait data according to Hamon and Noirot (1990) https://www.documentation.ird.fr/hor/fdi:36506, Noirot et al. (1996) doi:10.2307/2527837 and Noirot et al. (2003) https://www.documentation.ird.fr/hor/fdi:010031886.

See also

Author

Maintainer: J. Aravind j.aravind@icar.org.in (ORCID)

Authors:

J. Aravind j.aravind@icar.org.in (ORCID)

Anju Mahendru Singh anju.singh@icar.org.in (ORCID)

Other contributors:

ICAR-NBGPR (ROR) (url: https://nbpgr.org.in/) [copyright holder]